+++ src/IRutils.h
@@ -30,7 +30,7 @@ String resultToHumanReadableBasic(const decode_results * const results);
 String resultToHexidecimal(const decode_results * const result);
 bool hasACState(const decode_type_t protocol);
 uint16_t getCorrectedRawLength(const decode_results * const results);
-uint16_t *resultToRawArray(const decode_results * const decode);
+uint32_t *resultToRawArray(const decode_results * const decode);
 uint8_t sumBytes(const uint8_t * const start, const uint16_t length,
                  const uint8_t init = 0);
 uint8_t xorBytes(const uint8_t * const start, const uint16_t length,
